Throw on missing function arguments (#94), resolves #93
In #93 I noted that JSEP ignored missing function arguments. This code will throw in that case (see test for examples).

What I did not realize at the time was that arrays with missing elements were also broken, in the same code path. This PR fixes that as well, so 

`[1,,2]` returns `[1, null, 2]` (`null` is what esprima inserts in this case; I would prefer `undefined`, but I'm trying to keep it consistent). 

Finally I noticed that esprima is actually incorrect in the case of a trailing, missing array entry, so 
`[1,,]` in esprima returns `[1,null]` (a single trailing null) instead of the expected `[1,null,null]`. This could be fixed but I opted here to just replicate the esprima behavior. (Also FWIW those two arrays are functionally equivalent).