fix: firefox compatibility
When copying from the static class methods to the jsep function, don't rely on the ordering of keys. Instead, check for properties that are allowed to be set (![prototype, name, length]), by checking if they already exist on jsep. This fixes the "length" is readonly error in firefox (but works in Chrome/Node)

Fixes #178